MNIST with architecture 64-8-512
We train 100 CNN's of type M(64,8,215)* on the 60,000 traning examples of MNIST for 20 epochs with a batch size of 128 to a test accuracy of over 99%. We then collect the weights of the fitlers in the first convolutional layer of these models. After that we use mapper to produce a mapper graph.